backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.4.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
charset-normalizer==2.0.9
click==8.0.3
ecdsa==0.17.0
fastapi==0.70.1
frozenlist==1.2.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
multidict==5.2.0
psycopg2==2.9.2
pyasn1==0.4.8
pycodestyle==2.8.0
pydantic==1.8.2
python-jose==3.3.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.16.0
toml==0.10.2
typing-extensions==4.0.1
uvicorn==0.16.0
yarl==1.7.2
